refactor(callback): seperate into CallbackSection and CallbackPage to correctly suspend searchParams
